+++ src/shared_math/stark/brainfuck/stark.rs
@@ -27,7 +27,7 @@ use std::convert::TryInto;
 use std::error::Error;
 use std::rc::Rc;
 
-pub const EXTENSION_CHALLENGE_COUNT: u16 = 11;
+pub const EXTENSION_CHALLENGE_COUNT: usize = 11;
 pub const PERMUTATION_ARGUMENTS_COUNT: usize = 2;
 pub const TERMINAL_COUNT: usize = 5;
 
@@ -52,11 +52,11 @@ pub struct Stark {
 impl Stark {
     // TODO: Change this to use Rescue prime instead of Vec<u8>/Blake3
     // TODO: Use simple_hasher's get_n_hash_rounds() instead.
-    fn sample_weights(number: u16, seed: Vec<u8>) -> Vec<XFieldElement> {
+    fn sample_weights(number: u8, seed: Vec<u8>) -> Vec<XFieldElement> {
         let mut challenges: Vec<XFieldElement> = vec![];
         for i in 0..number {
             let mut mutated_challenge_seed = seed.clone();
-            mutated_challenge_seed[0] = ((mutated_challenge_seed[0] as u16 + i) % 256) as u8;
+            mutated_challenge_seed[0] = ((mutated_challenge_seed[0] as u16 + i as u16) % 256) as u8;
             // This is wrong:
             challenges.push(XFieldElement::ring_zero().from_vecu8(mutated_challenge_seed));
         }
@@ -308,10 +308,12 @@ impl Stark {
         // let challenges = self.sample_weights(EXTENSION_CHALLENGE_COUNT, proof_stream.prover_fiat_shamir());
         // TODO: REPLACE THIS WITH RescuePrime/B field elements. The type of `challenges`
         // must not change though, it should remain `Vec<XFieldElement>`.
-        let challenges: [XFieldElement; EXTENSION_CHALLENGE_COUNT as usize] =
-            Self::sample_weights(EXTENSION_CHALLENGE_COUNT, proof_stream.prover_fiat_shamir())
-                .try_into()
-                .unwrap();
+        let challenges: [XFieldElement; EXTENSION_CHALLENGE_COUNT] = Self::sample_weights(
+            EXTENSION_CHALLENGE_COUNT as u8,
+            proof_stream.prover_fiat_shamir(),
+        )
+        .try_into()
+        .unwrap();
 
         let initials: [XFieldElement; PERMUTATION_ARGUMENTS_COUNT] =
             XFieldElement::random_elements(PERMUTATION_ARGUMENTS_COUNT, &mut rng)
